--- a/bin/norma.js
+++ b/bin/norma.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 "use strict";
-require("time-require");
 require("coffee-script/register");
 var Liftoff = require("liftoff");
 var Flags = require("minimist")(process.argv.slice(2));