--- a/lib/index.js
+++ b/lib/index.js
@@ -1,20 +1,3 @@
-//
-// require("coffee-script/register");
-//
-// // only create one instance
-// if (!Norma) {
-//   var Norma = require("./norma");
-//   require("./libraries")(Norma);
-//   // global namespace
-//   GLOBAL.Norma = Norma
-//
-//   // 0.9.7 poor decision support
-//   process.Norma = Norma
-// }
-//
-//
-// module.exports = Norma
-
 
 require("coffee-script/register");