--- a/lib/utilities/manage-dependencies.coffee
+++ b/lib/utilities/manage-dependencies.coffee
@@ -1,16 +1,20 @@
 Path = require "path"
 Fs = require "fs"
-Semver = require "semver"
+
 Q = require "kew"
 _ = require "underscore"
-# Exec = require("child_process").exec
-Spawn = require "./promise-spawn"
 
 
-Norma = require "./../norma"
-MapTree = require("./directory-tools").mapTree
 
 module.exports = (tasks, cwd, flush) ->
+
+  Semver = require "semver"
+
+
+
+  Norma = require "./../norma"
+  Spawn = require "./promise-spawn"
+  MapTree = require("./directory-tools").mapTree
 
   # create the deferred
   loaded = Q.defer()