--- a/package.json
+++ b/package.json
@@ -25,7 +25,6 @@
     "pipeline"
   ],
   "dependencies": {
-    "anymatch": "^1.3.0",
     "bcrypt": "^0.8.3",
     "chalk": "^0.5.0",
     "chokidar": "^1.0.3",
@@ -47,9 +46,10 @@
     "prettyjson": "^1.1.1",
     "rimraf": "^2.2.8",
     "semver": "^4.1.0",
+    "time-require": "^0.1.2",
     "underscore": "^1.7.0",
     "user-home": "^1.1.0",
-    "vinyl-fs": "^1.0.0"
+    "vinyl-fs": "git+https://github.com/normajs/vinyl-fs.git"
   },
   "devDependencies": {
     "chai": "^2.0.0",