--- a/package.json
+++ b/package.json
@@ -46,7 +46,6 @@
     "prettyjson": "^1.1.1",
     "rimraf": "^2.2.8",
     "semver": "^4.1.0",
-    "time-require": "^0.1.2",
     "underscore": "^1.7.0",
     "user-home": "^1.1.0",
     "vinyl-fs": "git+https://github.com/normajs/vinyl-fs.git"