--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "normajs",
-  "version": "1.6.2",
+  "version": "1.6.3",
   "description": "Auto Discover Build Utility",
   "repository": {
     "type": "git",