--- a/test/fixtures/package.json
+++ b/test/fixtures/package.json
@@ -7,9 +7,5 @@
     "test": "node ./testing-scripts/index.js"
   },
   "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "norma-copy": "^0.4.2"
-  },
-  "devDependencies": {}
+  "license": "MIT"
 }