--- a/test/fixtures/package.json
+++ b/test/fixtures/package.json
@@ -1,11 +1,13 @@
 {
   "name": "test-project",
-  "version": "0.0.0",
+  "version": "0.1.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
     "test": "node ./testing-scripts/index.js"
   },
   "author": "",
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "norma-copy": "^0.4.2"
+  }
 }