… test/fixtures/images/test.html
… test/fixtures/lib/test.js
… test/fixtures/out/images/test.html