In order for Norma to be picked up by mainstream dev teams, we need to make sure she is as fast as possible. This PR is the first stab at that targeted on heavy and unnecessary requires.

# Benchmarks

## Norma v1.6.2
*Require of bare project (no tasks)*
```bash
Start time: (2015-06-23 11:55:42 UTC) [treshold=1%,sorted]
# [order]  module                                                                                       time  %
 1 [1087]  ../lib/index (../../../normajs/norma/lib/index.js)                                           1.5s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 58%
 2  [411]  ./norma (../../../normajs/norma/lib/norma.coffee)                                           298ms  ▇▇▇▇▇▇▇▇▇▇▇ 11%
 3  [781]  /Users/james.baxley/Products/normajs/norma/l...../normajs/norma/lib/methods/create.coffee)  235ms  ▇▇▇▇▇▇▇▇▇ 9%
 4  [771]  ./init (../../../normajs/norma/lib/methods/init.coffee)                                     208ms  ▇▇▇▇▇▇▇▇ 8%
 5  [352]  vinyl-fs (../../../normajs/norma/node_modules/vinyl-fs/index.js)                            160ms  ▇▇▇▇▇▇ 6%
 6  [757]  inquirer (../../../normajs/norma/node_modules/inquirer/lib/inquirer.js)                     145ms  ▇▇▇▇▇▇ 6%
 7 [1473]  /Users/james.baxley/Products/normajs/norma/n...e_modules/npm-registry-client/lib/fetch.js)  137ms  ▇▇▇▇▇ 5%
 8  [329]  ./lib/src (../../../normajs/norma/node_modules/vinyl-fs/lib/src/index.js)                   137ms  ▇▇▇▇▇ 5%
 9 [1471]  request (../../../normajs/norma/node_modules/npm/node_modules/request/index.js)             135ms  ▇▇▇▇▇ 5%
10  [926]  /Users/james.baxley/Products/normajs/norma/l..../normajs/norma/lib/methods/install.coffee)  131ms  ▇▇▇▇▇ 5%
11 [1085]  /Users/james.baxley/Products/normajs/norma/l.../../normajs/norma/lib/methods/watch.coffee)  129ms  ▇▇▇▇▇ 5%
12 [1238]  ./launcher (../../../normajs/norma/bin/launcher.coffee)                                     129ms  ▇▇▇▇▇ 5%
13  [701]  ./prompts/list (../../../normajs/norma/node_modules/inquirer/lib/prompts/list.js)           118ms  ▇▇▇▇▇ 5%
14 [1237]  ./auto-update (../../../normajs/norma/bin/auto-update.coffee)                               113ms  ▇▇▇▇ 4%
15 [1470]  ./request (../../../normajs/norma/node_modules/npm/node_modules/request/request.js)         110ms  ▇▇▇▇ 4%
16  [699]  ./base (../../../normajs/norma/node_modules/inquirer/lib/prompts/base.js)                    95ms  ▇▇▇▇ 4%
17 [1084]  ./../utilities/watch (../../../normajs/norma/lib/utilities/watch.coffee)                     93ms  ▇▇▇▇ 4%
18 [1230]  npm (../../../normajs/norma/node_modules/npm/lib/npm.js)                                     93ms  ▇▇▇▇ 4%
19  [492]  ./utilities/read-settings (../../../normajs/norma/lib/utilities/read-settings.coffee)        85ms  ▇▇▇ 3%
20  [684]  cli-color (../../../normajs/norma/node_modules/inquirer/node_modules/cli-color/index.js)     83ms  ▇▇▇ 3%
21  [918]  github-download (../../../normajs/norma/node...les/github-download/lib/github-download.js)   78ms  ▇▇▇ 3%
22  [581]  /Users/james.baxley/Products/normajs/norma/l.../../normajs/norma/lib/methods/build.coffee)   75ms  ▇▇▇ 3%
23  [541]  ./utilities/register-packages (../../../norm...rma/lib/utilities/register-packages.coffee)   74ms  ▇▇▇ 3%
24 [1081]  chokidar (../../../normajs/norma/node_modules/chokidar/index.js)                             73ms  ▇▇▇ 3%
25  [409]  ./utilities/directory-tools (../../../normajs/norma/lib/utilities/directory-tools.coffee)    71ms  ▇▇▇ 3%
26   [95]  liftoff (../../../normajs/norma/node_modules/liftoff/index.js)                               62ms  ▇▇▇ 2%
27   [28]  coffee-script/register (../../../normajs/norma/node_modules/coffee-script/register.js)       57ms  ▇▇ 2%
28  [978]  /Users/james.baxley/Products/normajs/norma/l..../../normajs/norma/lib/methods/test.coffee)   56ms  ▇▇ 2%
29   [27]  ./lib/coffee-script/register (../../../norma...offee-script/lib/coffee-script/register.js)   56ms  ▇▇ 2%
30  [515]  ./utilities/manage-dependencies (../../../no...a/lib/utilities/manage-dependencies.coffee)   55ms  ▇▇ 2%
31 [1046]  anymatch (../../../normajs/norma/node_modules/anymatch/index.js)                             48ms  ▇▇ 2%
32   [22]  ./coffee-script (../../../normajs/norma/node...-script/lib/coffee-script/coffee-script.js)   48ms  ▇▇ 2%
33 [1166]  ./config/core.js (../../../normajs/norma/node_modules/npm/lib/config/core.js)                47ms  ▇▇ 2%
34 [1044]  micromatch (../../../normajs/norma/node_modu.../anymatch/node_modules/micromatch/index.js)   46ms  ▇▇ 2%
35  [665]  memoizee (../../../normajs/norma/node_module...s/cli-color/node_modules/memoizee/index.js)   46ms  ▇▇ 2%
36  [210]  glob-stream (../../../normajs/norma/node_mod...vinyl-fs/node_modules/glob-stream/index.js)   42ms  ▇▇ 2%
37  [664]  ./plain (../../../normajs/norma/node_modules...s/cli-color/node_modules/memoizee/plain.js)   41ms  ▇▇ 2%
38  [430]  ./utilities/prompt (../../../normajs/norma/lib/utilities/prompt.coffee)                      40ms  ▇▇ 2%
39  [845]  request (../../../normajs/norma/node_modules...thub-download/node_modules/request/main.js)   39ms  ▇▇ 2%
40  [662]  ./lib/configure-map (../../../normajs/norma/...node_modules/memoizee/lib/configure-map.js)   37ms  ▇▇ 1%
41   [89]  ./lib/find_config (../../../normajs/norma/node_modules/liftoff/lib/find_config.js)           35ms  ▇▇ 1%
42   [87]  findup-sync (../../../normajs/norma/node_mod...ode_modules/findup-sync/lib/findup-sync.js)   34ms  ▇▇ 1%
43   [88]  ./file_search (../../../normajs/norma/node_modules/liftoff/lib/file_search.js)               34ms  ▇▇ 1%
44  [768]  ./../utilities/scaffold (../../../normajs/norma/lib/utilities/scaffold.coffee)               33ms  ▇▇ 1%
45  [534]  ./package-lookup (../../../normajs/norma/lib/utilities/package-lookup.coffee)                33ms  ▇▇ 1%
46 [1041]  ./lib/expand (../../../normajs/norma/node_mo...atch/node_modules/micromatch/lib/expand.js)   32ms  ▇▇ 1%
47  [499]  ./utilities/read-config (../../../normajs/norma/lib/utilities/read-config.coffee)            32ms  ▇▇ 1%
48 [1526]  /Users/james.baxley/Products/normajs/norma/n...modules/npm-registry-client/lib/publish.js)   32ms  ▇▇ 1%
49 [1597]  /Users/james.baxley/Products/normajs/norma/n...normajs/norma/node_modules/npm/lib/view.js)   30ms  ▇▇ 1%
50   [86]  glob (../../../normajs/norma/node_modules/li...ules/findup-sync/node_modules/glob/glob.js)   29ms  ▇▇ 1%
51  [965]  /Users/james.baxley/Products/normajs/norma/l.../normajs/norma/lib/methods/settings.coffee)   28ms  ▇ 1%
52 [1227]  ./cache/caching-client.js (../../../normajs/...de_modules/npm/lib/cache/caching-client.js)   28ms  ▇ 1%
53 [1040]  ./glob (../../../normajs/norma/node_modules/anymatch/node_modules/micromatch/lib/glob.js)    27ms  ▇ 1%
54  [136]  through2 (../../../normajs/norma/node_modules/vinyl-fs/node_modules/through2/through2.js)    26ms  ▇ 1%
Total require(): 1600
Total time: 2.6s
```

### Total time to require: _2.6s_



## Norma v.1.6.3 (current branch)
*Require of bare project (no tasks)*

```bash
Start time: (2015-06-23 11:53:53 UTC) [treshold=1%,sorted]
# [order]  module                                                                                       time  %
  1 [630]  ../lib/index (../../../normajs/norma/lib/index.js)                                          904ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 77%
  2 [409]  ./norma (../../../normajs/norma/lib/norma.coffee)                                           275ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 23%
  3 [351]  vinyl-fs (../../../normajs/norma/node_modules/vinyl-fs/index.js)                            143ms  ▇▇▇▇▇▇▇▇▇▇▇ 12%
  4 [328]  ./lib/src (../../../normajs/norma/node_modules/vinyl-fs/lib/src/index.js)                   126ms  ▇▇▇▇▇▇▇▇▇▇ 11%
  5 [484]  ./utilities/read-settings (../../../normajs/norma/lib/utilities/read-settings.coffee)        67ms  ▇▇▇▇▇▇ 6%
  6 [407]  ./utilities/directory-tools (../../../normajs/norma/lib/utilities/directory-tools.coffee)    63ms  ▇▇▇▇▇ 5%
  7  [28]  coffee-script/register (../../../normajs/norma/node_modules/coffee-script/register.js)       46ms  ▇▇▇▇ 4%
  8 [209]  glob-stream (../../../normajs/norma/node_mod...vinyl-fs/node_modules/glob-stream/index.js)   45ms  ▇▇▇▇ 4%
  9  [27]  ./lib/coffee-script/register (../../../norma...offee-script/lib/coffee-script/register.js)   45ms  ▇▇▇▇ 4%
 10 [588]  ./methods/install.coffee (../../../normajs/norma/lib/methods/install.coffee)                 43ms  ▇▇▇▇ 4%
 11  [95]  liftoff (../../../normajs/norma/node_modules/liftoff/index.js)                               41ms  ▇▇▇▇ 3%
 12 [671]  ./package-lookup (../../../normajs/norma/lib/utilities/package-lookup.coffee)                40ms  ▇▇▇▇ 3%
 13  [22]  ./coffee-script (../../../normajs/norma/node...-script/lib/coffee-script/coffee-script.js)   40ms  ▇▇▇▇ 3%
 14 [428]  ./utilities/prompt (../../../normajs/norma/lib/utilities/prompt.coffee)                      37ms  ▇▇▇ 3%
 15 [499]  ./utilities/manage-dependencies (../../../no...a/lib/utilities/manage-dependencies.coffee)   36ms  ▇▇▇ 3%
 16 [618]  ./methods/create.coffee (../../../normajs/norma/lib/methods/create.coffee)                   35ms  ▇▇▇ 3%
 17 [552]  ./methods/test.coffee (../../../normajs/norma/lib/methods/test.coffee)                       34ms  ▇▇▇ 3%
 18 [563]  ./methods/settings.coffee (../../../normajs/norma/lib/methods/settings.coffee)               32ms  ▇▇▇ 3%
 19 [536]  ./methods/watch.coffee (../../../normajs/norma/lib/methods/watch.coffee)                     29ms  ▇▇▇ 2%
 20 [677]  ./auto-discover (../../../normajs/norma/lib/utilities/auto-discover.coffee)                  29ms  ▇▇▇ 2%
 21  [88]  ./file_search (../../../normajs/norma/node_modules/liftoff/lib/file_search.js)               28ms  ▇▇▇ 2%
 22  [89]  ./lib/find_config (../../../normajs/norma/node_modules/liftoff/lib/find_config.js)           28ms  ▇▇▇ 2%
 23 [629]  ./methods/build.coffee (../../../normajs/norma/lib/methods/build.coffee)                     28ms  ▇▇▇ 2%
 24  [87]  findup-sync (../../../normajs/norma/node_mod...ode_modules/findup-sync/lib/findup-sync.js)   27ms  ▇▇▇ 2%
 25 [643]  ./launcher (../../../normajs/norma/bin/launcher.coffee)                                      27ms  ▇▇▇ 2%
 26  [86]  glob (../../../normajs/norma/node_modules/li...ules/findup-sync/node_modules/glob/glob.js)   26ms  ▇▇ 2%
 27 [566]  ./methods/search.coffee (../../../normajs/norma/lib/methods/search.coffee)                   26ms  ▇▇ 2%
 28 [491]  ./utilities/read-config (../../../normajs/norma/lib/utilities/read-config.coffee)            24ms  ▇▇ 2%
 29 [616]  ./../utilities/package (../../../normajs/norma/lib/utilities/package.coffee)                 24ms  ▇▇ 2%
 30 [573]  ./methods/remove.coffee (../../../normajs/norma/lib/methods/remove.coffee)                   23ms  ▇▇ 2%
 31 [568]  ./methods/run.coffee (../../../normajs/norma/lib/methods/run.coffee)                         22ms  ▇▇ 2%
 32 [415]  ./libraries (../../../normajs/norma/lib/libraries.coffee)                                    20ms  ▇▇ 2%
 33  [14]  ./parser (../../../normajs/norma/node_modules/coffee-script/lib/coffee-script/parser.js)     20ms  ▇▇ 2%
 34 [592]  ./methods/init.coffee (../../../normajs/norma/lib/methods/init.coffee)                       19ms  ▇▇ 2%
 35 [599]  ./methods/help.coffee (../../../normajs/norma/lib/methods/help.coffee)                       19ms  ▇▇ 2%
 36 [482]  cson (../../../normajs/norma/node_modules/cson/out/lib/cson.js)                              18ms  ▇▇ 2%
 37 [494]  ./utilities/orchestration (../../../normajs/norma/lib/utilities/orchestration.coffee)        18ms  ▇▇ 2%
 38 [266]  duplexify (../../../normajs/norma/node_modules/vinyl-fs/node_modules/duplexify/index.js)     18ms  ▇▇ 2%
 39 [135]  through2 (../../../normajs/norma/node_modules/vinyl-fs/node_modules/through2/through2.js)    17ms  ▇▇ 1%
 40 [306]  gulp-sourcemaps (../../../normajs/norma/node...l-fs/node_modules/gulp-sourcemaps/index.js)   17ms  ▇▇ 1%
 41 [431]  ./events/message (../../../normajs/norma/lib/events/message.coffee)                          16ms  ▇▇ 1%
 42 [598]  ./../logging/logger (../../../normajs/norma/lib/logging/logger.coffee)                       15ms  ▇▇ 1%
 43 [539]  ./methods/update.coffee (../../../normajs/norma/lib/methods/update.coffee)                   15ms  ▇▇ 1%
 44 [642]  ./auto-update (../../../normajs/norma/bin/auto-update.coffee)                                15ms  ▇▇ 1%
 45 [404]  fs-extra (../../../normajs/norma/node_modules/fs-extra/lib/index.js)                         15ms  ▇▇ 1%
 46 [504]  ./utilities/register-packages (../../../norm...rma/lib/utilities/register-packages.coffee)   14ms  ▇▇ 1%
 47 [261]  readable-stream (../../../normajs/norma/node...y/node_modules/readable-stream/readable.js)   14ms  ▇▇ 1%
 48 [551]  ./../utilities/run-script (../../../normajs/norma/lib/utilities/run-script.coffee)           13ms  ▇ 1%
 49 [131]  ./lib/_stream_transform.js (../../../normajs...s/readable-stream/lib/_stream_transform.js)   13ms  ▇ 1%
 50 [132]  readable-stream/transform (../../../normajs/.../node_modules/readable-stream/transform.js)   13ms  ▇ 1%
 51 [203]  glob (../../../normajs/norma/node_modules/vi...ules/glob-stream/node_modules/glob/glob.js)   13ms  ▇ 1%
 52 [459]  nconf (../../../normajs/norma/node_modules/nconf/lib/nconf.js)                               13ms  ▇ 1%
 53 [168]  ordered-read-streams (../../../normajs/norma...node_modules/ordered-read-streams/index.js)   13ms  ▇ 1%
 54 [128]  ./_stream_duplex (../../../normajs/norma/nod...ules/readable-stream/lib/_stream_duplex.js)   12ms  ▇ 1%
 55 [164]  readable-stream (../../../normajs/norma/node...s/node_modules/readable-stream/readable.js)   12ms  ▇ 1%
 56 [651]  ./promise-spawn (../../../normajs/norma/lib/utilities/promise-spawn.coffee)                  12ms  ▇ 1%
 57 [346]  ./lib/dest (../../../normajs/norma/node_modules/vinyl-fs/lib/dest/index.js)                  12ms  ▇ 1%
 58 [628]  ./../utilities/generate-task-list (../../../...ma/lib/utilities/generate-task-list.coffee)   12ms  ▇ 1%
Total require(): 680
Total time: 1.2s
```

### Total time to require: _1.2s_

# 54% reduction of load times for base libraries